fixed merge conflicts and e2e tests
