+++ apps/ods-api/src/surveyResponse/surveyResponse.service.ts
@@ -1,6 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { SurveyResponse, Prisma, Survey, Answer } from '.prisma/ods/client';
 import { PrismaService } from '../prisma/prisma.service';
+// eslint-disable-next-line no-restricted-imports
+import { ResponseCount } from '@odst/types/ods';
 
 @Injectable()
 export class SurveyResponseService {
@@ -75,10 +77,11 @@ export class SurveyResponseService {
   }
 
   // Get the string IDs of all the issues that are unresolved that the commander
+  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
   async getSurveyResponseData(
     surveyResponseWhereUniqueInput: Prisma.SurveyResponseWhereUniqueInput
   ) {
-    return await this.prisma.surveyResponse.findUnique({
+    return this.prisma.surveyResponse.findUnique({
       where: {
         id: surveyResponseWhereUniqueInput.id,
       },
@@ -92,7 +95,9 @@ export class SurveyResponseService {
     });
   }
 
-  async create(data: Prisma.SurveyResponseCreateInput) {
+  async create(
+    data: Prisma.SurveyResponseCreateInput
+  ): Promise<SurveyResponse> {
     return this.prisma.surveyResponse.create({
       data,
     });
@@ -123,7 +128,7 @@ export class SurveyResponseService {
 
   // TODO: Optimize at a later date, so we don't go back and forth to the server
 
-  async countResponses() {
+  async countResponses(): Promise<ResponseCount> {
     //TODO: Promise a number array and remove awaits
 
     const unresolvedCount = await this.prisma.surveyResponse.count({
