+++ libs/types/ods/src/index.ts
@@ -11,7 +11,6 @@ export * from './entity/user.entity';
 export * from './input/answer.create.input';
 export * from './input/answer.update.input';
 export * from './input/answer.unique.input';
-export * from './input/answer.where.input';
 export * from './input/org.create.input';
 export * from './input/org.update.input';
 export * from './input/org.unique.input';
@@ -21,17 +20,10 @@ export * from './input/question.unique.input';
 export * from './input/survey.create.input';
 export * from './input/survey.update.input';
 export * from './input/survey.unique.input';
-export * from './input/survey.where.input';
 export * from './input/surveyResponse.create.input';
 export * from './input/surveyResponse.update.input';
 export * from './input/surveyResponse.unique.input';
-export * from './input/surveyResponse.where.input';
 export * from './input/user.create.input';
 export * from './input/user.update.input';
 export * from './input/user.unique.input';
 //#endregion
-
-//#region reusables
-// TODO move these to a common location
-export * from './reusables/reusables.input';
-//#endregion reusables
